fixed timeline to dynamically render items
