+++ app/src/main/java/com/tripper/TimeLineItem.java
@@ -1,23 +1,17 @@
 package com.tripper;
 
 public class TimeLineItem {
-    public int position;
     public int imageName;
     public String location;
     public String timeOfDay;
     public String month;
     public String day;
 
-    public TimeLineItem(int position, int imageName, String location, String timeOfDay, String month, String day) {
-        this.position = position;
+    public TimeLineItem(int imageName, String location, String timeOfDay, String month, String day) {
         this.imageName = imageName;
         this.location = location;
         this.timeOfDay = timeOfDay;
         this.month = month;
         this.day = day;
-    }
-
-    public int getPosition(){
-        return position;
     }
 }
